src/requirements.txt
flask==2.2.5
pika==1.3.2
pytesseract==0.3.10
psycopg2-binary==2.9.6
pdf2image==1.16.3
Pillow==9.5.0
